// Capstone_Project/Frontend/PropZone/src/app/models/basic-property-search.model.ts
export interface BasicPropertySearchModel {
  locality?: string;
  city?: string;
  listingPurpose?: 'Sale' | 'Rent'; 
  propertyTypes?: string[]; 
  listerTypes?: string[]; 
  minPrice?: number;
  maxPrice?: number;
  keyword?: string;
  minArea?: number;
  maxArea?: number;
  hasImages?: boolean;
  sortBy?: 'CreatedAt' | 'Price';
  ascending?: boolean;
  status?: 'Available' | 'Sold' | 'Rented';

  // admin
  listerId?: string;
  statuses?: string[];

}